+++ pkg/mount/mount.go
@@ -66,3 +66,22 @@ func MountXFSDevice(device, target string, flags []string) error {
 	klog.V(3).InfoS("mounting device", "device", device, "target", target)
 	return SafeMount(device, target, "xfs", flags, MountOptPrjQuota)
 }
+
+func ValidDirectPVMountOpts(deviceMountOpts []string) bool {
+	expectedMountOpts := []string{
+		MountOptPrjQuota,
+	}
+	for _, expectedMountOpt := range expectedMountOpts {
+		foundExpectedOpt := false
+		for _, deviceMountOpt := range deviceMountOpts {
+			if deviceMountOpt == expectedMountOpt {
+				foundExpectedOpt = true
+				break
+			}
+		}
+		if !foundExpectedOpt {
+			return false
+		}
+	}
+	return true
+}
