+++ pkg/uevent/utils_test.go
@@ -1007,7 +1007,7 @@ func TestValidateDevice(t1 *testing.T) {
 	}
 
 	for i, testCase := range testCases {
-		if testCase.expectedResult != ValidateDevice(testCase.device, testCase.drives) {
+		if testCase.expectedResult != ValidateDevice(testCase.device, testCase.drives[0]) {
 			t1.Errorf("Test case %d: Expected result = (%v) got: %v", i, testCase.expectedResult, !testCase.expectedResult)
 		}
 	}
